mlops/pipelines/digital_twin_resilience/requirements.txt
boto3
sagemaker>=2.130,<3
numpy
pandas
scikit-learn
joblib
python-dotenv
pytz
openai
